Refactor governance lock for code review
