+++ shardingsphere-governance/shardingsphere-governance-core/src/main/java/org/apache/shardingsphere/governance/core/facade/listener/GovernanceListenerManager.java
@@ -18,7 +18,6 @@
 package org.apache.shardingsphere.governance.core.facade.listener;
 
 import org.apache.shardingsphere.governance.core.config.listener.ConfigurationListenerManager;
-import org.apache.shardingsphere.governance.core.lock.listener.LockListenerManager;
 import org.apache.shardingsphere.governance.core.registry.listener.RegistryListenerManager;
 import org.apache.shardingsphere.governance.repository.api.ConfigurationRepository;
 import org.apache.shardingsphere.governance.repository.api.RegistryRepository;
@@ -34,12 +33,9 @@ public final class GovernanceListenerManager {
     
     private final RegistryListenerManager registryListenerManager;
     
-    private final LockListenerManager lockListenerManager;
-    
     public GovernanceListenerManager(final RegistryRepository registryRepository, final ConfigurationRepository configurationRepository, final Collection<String> schemaNames) {
         configurationListenerManager = new ConfigurationListenerManager(configurationRepository, schemaNames);
         registryListenerManager = new RegistryListenerManager(registryRepository, schemaNames);
-        lockListenerManager = new LockListenerManager(registryRepository);
     }
     
     /**
@@ -48,6 +44,5 @@ public final class GovernanceListenerManager {
     public void init() {
         configurationListenerManager.initListeners();
         registryListenerManager.initListeners();
-        lockListenerManager.initListeners();
     }
 }
