--- shardingsphere-governance/shardingsphere-governance-core/src/main/java/org/apache/shardingsphere/governance/core/lock/listener/LockListenerManager.java
@@ -1,40 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shardingsphere.governance.core.lock.listener;
-
-import org.apache.shardingsphere.governance.repository.api.RegistryRepository;
-import org.apache.shardingsphere.governance.repository.api.listener.DataChangedEvent.Type;
-
-/**
- * Lock listener manager.
- */
-public final class LockListenerManager {
-    
-    private final GlobalLockChangedListener globalLockChangedListener;
-    
-    public LockListenerManager(final RegistryRepository registryRepository) {
-        globalLockChangedListener = new GlobalLockChangedListener(registryRepository);
-    }
-    
-    /**
-     * Initialize all distributed lock listeners.
-     */
-    public void initListeners() {
-        globalLockChangedListener.watch(Type.ADDED);
-    }
-}
